+++ src/app/groups/[id]/page.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { useParams } from 'next/navigation';
 import { mockGroups } from '@/mocks/data';
 import { mockPosts } from '@/mocks/posts';
-import BoardHeader from '@/components/features/post/BoardHeader';
+import BoardHeader from '@/app/groups/components/BoardHeader';
 import Sidebar from '@/app/groups/components/Sidebar';
 import PostList from '@/app/groups/components/PostList';
 import Calendar from '@/app/groups/components/Calendar';
